完成编辑媒体、收藏夹功能

+++ lib/service/media/article_service.dart
@@ -15,9 +15,20 @@ class ArticleService {
     return article;
   }
 
-  static Future<void> deleteArticle(
-    String articleId,
+  static Future<void> updateArticleData(
+    String mediaId,
+    String? title,
+    String? introduction,
+    String? content,
+    String? coverUrl,
   ) async {
+    await ArticleApi.updateArticleData(mediaId, title, introduction, content, coverUrl);
+  }
+
+  static Future<void> deleteArticle(
+    String? articleId,
+  ) async {
+    if(articleId==null) return;
     await ArticleApi.deleteUserArticleById(articleId);
   }
